Minor_Project_UPI_Call_Fraud_Detection/app/requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
tensorflow
imblearn
plotly
joblib
pyyaml
